refactor: centralize startup command bootstrap
